+++ packages/svelte/src/internal/client/dom/blocks/html.js
@@ -1,29 +1,22 @@
+import { derived } from '../../reactivity/deriveds.js';
 import { render_effect } from '../../reactivity/effects.js';
+import { get } from '../../runtime.js';
 import { reconcile_html, remove } from '../reconciler.js';
 
 /**
- * @param {Element | Text | Comment} dom
+ * @param {Element | Text | Comment} anchor
  * @param {() => string} get_value
  * @param {boolean} svg
  * @returns {void}
  */
-export function html(dom, get_value, svg) {
-	/** @type {import('#client').Dom} */
-	let html_dom;
+export function html(anchor, get_value, svg) {
+	let value = derived(get_value);
 
-	/** @type {string} */
-	let value;
+	render_effect(() => {
+		var dom = reconcile_html(anchor, get(value), svg);
 
-	const effect = render_effect(() => {
-		if (value !== (value = get_value())) {
-			if (html_dom) remove(html_dom);
-			html_dom = reconcile_html(dom, value, svg);
+		if (dom) {
+			return () => remove(dom);
 		}
 	});
-
-	effect.ondestroy = () => {
-		if (html_dom) {
-			remove(html_dom);
-		}
-	};
 }