+++ packages/svelte/src/internal/client/dom/template.js
@@ -1,4 +1,4 @@
-import { hydrate_nodes, hydrate_block_anchor, hydrating } from './hydration.js';
+import { hydrate_nodes, hydrating } from './hydration.js';
 import { child, clone_node, empty } from './operations.js';
 import {
 	create_fragment_from_html,
@@ -83,18 +83,12 @@ export function svg_template_with_script(svg, return_fragment) {
 /**
  * @param {boolean} is_fragment
  * @param {boolean} use_clone_node
- * @param {null | Text | Comment | Element} anchor
  * @param {() => Node} [template_element_fn]
  * @returns {Element | DocumentFragment | Node[]}
  */
 /*#__NO_SIDE_EFFECTS__*/
-function open_template(is_fragment, use_clone_node, anchor, template_element_fn) {
+function open_template(is_fragment, use_clone_node, template_element_fn) {
 	if (hydrating) {
-		if (anchor !== null) {
-			// TODO why is this sometimes null and sometimes not? needs clear documentation
-			hydrate_block_anchor(anchor);
-		}
-
 		return is_fragment ? hydrate_nodes : /** @type {Element} */ (hydrate_nodes[0]);
 	}
 
@@ -104,69 +98,46 @@ function open_template(is_fragment, use_clone_node, anchor, template_element_fn)
 }
 
 /**
- * @param {null | Text | Comment | Element} anchor
  * @param {() => Node} template_element_fn
  * @param {boolean} [use_clone_node]
  * @returns {Element}
  */
-export function open(anchor, template_element_fn, use_clone_node = true) {
-	return /** @type {Element} */ (open_template(false, use_clone_node, anchor, template_element_fn));
+export function open(template_element_fn, use_clone_node = true) {
+	return /** @type {Element} */ (open_template(false, use_clone_node, template_element_fn));
 }
 
 /**
- * @param {null | Text | Comment | Element} anchor
  * @param {() => Node} template_element_fn
  * @param {boolean} [use_clone_node]
  * @returns {Element | DocumentFragment | Node[]}
  */
-export function open_frag(anchor, template_element_fn, use_clone_node = true) {
-	return open_template(true, use_clone_node, anchor, template_element_fn);
+export function open_frag(template_element_fn, use_clone_node = true) {
+	return open_template(true, use_clone_node, template_element_fn);
 }
 
-const space_template = template(' ', false);
 const comment_template = template('<!>', true);
-
-/**
- * @param {Text | Comment | Element | null} anchor
- */
-/*#__NO_SIDE_EFFECTS__*/
-export function space_frag(anchor) {
-	/** @type {Node | null} */
-	var node = /** @type {any} */ (open(anchor, space_template));
-	// if an {expression} is empty during SSR, there might be no
-	// text node to hydrate (or an anchor comment is falsely detected instead)
-	//  — we must therefore create one
-	if (hydrating && node?.nodeType !== 3) {
-		node = empty();
-		// @ts-ignore in this case the anchor should always be a comment,
-		// if not something more fundamental is wrong and throwing here is better to bail out early
-		anchor.before(node);
-	}
-	return node;
-}
 
 /**
  * @param {Text | Comment | Element} anchor
  */
 /*#__NO_SIDE_EFFECTS__*/
-export function space(anchor) {
-	// if an {expression} is empty during SSR, there might be no
-	// text node to hydrate (or an anchor comment is falsely detected instead)
-	//  — we must therefore create one
-	if (hydrating && anchor.nodeType !== 3) {
-		const node = empty();
-		anchor.before(node);
-		return node;
+export function text(anchor) {
+	if (!hydrating) return empty();
+
+	var node = hydrate_nodes[0];
+
+	if (!node) {
+		// if an {expression} is empty during SSR, `hydrate_nodes` will be empty.
+		// we need to insert an empty text node
+		anchor.before((node = empty()));
 	}
-	return anchor;
+
+	return node;
 }
 
-/**
- * @param {null | Text | Comment | Element} anchor
- */
 /*#__NO_SIDE_EFFECTS__*/
-export function comment(anchor) {
-	return open_frag(anchor, comment_template);
+export function comment() {
+	return open_frag(comment_template);
 }
 
 /**
@@ -174,7 +145,7 @@ export function comment(anchor) {
  * and insert the elements into the dom (in client mode).
  * @param {import('#client').Dom} dom
  * @param {boolean} is_fragment
- * @param {null | Text | Comment | Element} anchor
+ * @param {Text | Comment | Element} anchor
  * @returns {import('#client').Dom}
  */
 function close_template(dom, is_fragment, anchor) {
@@ -187,10 +158,8 @@ function close_template(dom, is_fragment, anchor) {
 			current = /** @type {import('#client').Dom} */ ([.../** @type {Node} */ (dom).childNodes]);
 		}
 
-		if (anchor !== null) {
-			// TODO as with `open_template — why is this sometimes null and sometimes not?
-			insert(current, anchor);
-		}
+		// TODO ideally we'd do `anchor.before(dom)`, but that fails because `dom` can be an array of nodes in the SVG case
+		insert(current, anchor);
 	}
 
 	/** @type {import('#client').Effect} */ (current_effect).dom = current;
@@ -199,7 +168,7 @@ function close_template(dom, is_fragment, anchor) {
 }
 
 /**
- * @param {null | Text | Comment | Element} anchor
+ * @param {Text | Comment | Element} anchor
  * @param {Element | Text} dom
  */
 export function close(anchor, dom) {
@@ -207,7 +176,7 @@ export function close(anchor, dom) {
 }
 
 /**
- * @param {null | Text | Comment | Element} anchor
+ * @param {Text | Comment | Element} anchor
  * @param {Element | Text} dom
  */
 export function close_frag(anchor, dom) {