+++ packages/svelte/src/internal/client/runtime.js
@@ -20,8 +20,10 @@ import {
 	UNOWNED,
 	DESTROYED,
 	INERT,
-	MANAGED,
-	STATE_SYMBOL
+	BRANCH_EFFECT,
+	STATE_SYMBOL,
+	BLOCK_EFFECT,
+	ROOT_EFFECT
 } from './constants.js';
 import { flush_tasks } from './dom/task.js';
 import { add_owner } from './dev/ownership.js';
@@ -356,7 +358,9 @@ export function destroy_children(signal) {
 	if (signal.effects) {
 		for (var i = 0; i < signal.effects.length; i += 1) {
 			var effect = signal.effects[i];
-			if ((effect.f & MANAGED) === 0) {
+
+			// TODO ideally root effects would be parentless
+			if ((effect.f & ROOT_EFFECT) === 0) {
 				destroy_effect(effect);
 			}
 		}
@@ -376,7 +380,9 @@ export function destroy_children(signal) {
  * @returns {void}
  */
 export function execute_effect(effect) {
-	if ((effect.f & DESTROYED) !== 0) {
+	var flags = effect.f;
+
+	if ((flags & DESTROYED) !== 0) {
 		return;
 	}
 
@@ -391,7 +397,10 @@ export function execute_effect(effect) {
 	current_component_context = component_context;
 
 	try {
-		destroy_children(effect);
+		if ((flags & BLOCK_EFFECT) === 0) {
+			destroy_children(effect);
+		}
+
 		effect.teardown?.();
 		var teardown = execute_reaction_fn(effect);
 		effect.teardown = typeof teardown === 'function' ? teardown : null;
@@ -401,7 +410,7 @@ export function execute_effect(effect) {
 	}
 	const parent = effect.parent;
 
-	if ((effect.f & PRE_EFFECT) !== 0 && parent !== null) {
+	if ((flags & PRE_EFFECT) !== 0 && parent !== null) {
 		flush_local_pre_effects(parent);
 	}
 }
@@ -494,7 +503,6 @@ export function schedule_effect(signal) {
 			queueMicrotask(process_microtask);
 		}
 	}
-
 	let root = signal;
 
 	while (root !== null) {
@@ -502,10 +510,11 @@ export function schedule_effect(signal) {
 		if (parent === null) {
 			break;
 		}
-		if ((parent.f & CLEAN) !== 0) {
+		if ((parent.f & BRANCH_EFFECT) !== 0) {
+			if ((parent.f & CLEAN) === 0) {
+				return;
+			}
 			set_signal_status(parent, MAYBE_DIRTY);
-		} else {
-			return;
 		}
 		root = parent;
 	}
@@ -540,22 +549,22 @@ function collect_effects_recursively(
 	var pre = [];
 	var render = [];
 	var user = [];
-	var is_managed;
+	var is_branch;
 
 	for (i = 0; i < effects.length; i++) {
 		child = effects[i];
 		flags = child.f;
-		is_managed = flags & MANAGED;
+		is_branch = flags & BRANCH_EFFECT;
 		// TODO: put this in when we have proper working removal of managed effects from their parents
 		// if ((flags & CLEAN) !== 0) {
 		// 	continue;
 		// }
-		if (is_managed) {
+		if (is_branch) {
 			set_signal_status(child, CLEAN);
 		}
 
 		if ((flags & PRE_EFFECT) !== 0) {
-			if (is_managed) {
+			if (is_branch) {
 				if (shallow) {
 					return;
 				}
@@ -570,7 +579,7 @@ function collect_effects_recursively(
 				pre.push(child);
 			}
 		} else if ((flags & RENDER_EFFECT) !== 0) {
-			if (is_managed) {
+			if (is_branch) {
 				if (shallow) {
 					return;
 				}
@@ -585,7 +594,7 @@ function collect_effects_recursively(
 				render.push(child);
 			}
 		} else if ((flags & EFFECT) !== 0) {
-			if (is_managed) {
+			if (is_branch) {
 				if (shallow) {
 					return;
 				}
@@ -772,7 +781,11 @@ export function get(signal) {
 	}
 
 	// Register the dependency on the current reaction signal.
-	if (current_reaction !== null && (current_reaction.f & MANAGED) === 0 && !current_untracking) {
+	if (
+		current_reaction !== null &&
+		(current_reaction.f & BRANCH_EFFECT) === 0 &&
+		!current_untracking
+	) {
 		const unowned = (current_reaction.f & UNOWNED) !== 0;
 		const dependencies = current_reaction.deps;
 		if (
@@ -797,7 +810,7 @@ export function get(signal) {
 			current_untracked_writes !== null &&
 			current_effect !== null &&
 			(current_effect.f & CLEAN) !== 0 &&
-			(current_effect.f & MANAGED) === 0 &&
+			(current_effect.f & BRANCH_EFFECT) === 0 &&
 			current_untracked_writes.includes(signal)
 		) {
 			set_signal_status(current_effect, DIRTY);